Rename test and add helpers
